Refactor player data layer, migrate team IDs to Long, and implement jersey number validation
